First version of SKFS.
A tone of things are not probably handled (reader's dependencies are not yet
tracked etc ...), but this first version gives a good idea of what SKFS is.

One can create directories, and then attach (map) functions to those, whose
result is then put into anoter directory. The call looks like this:

myOutputDir = myInputDir.multiMap(context, "/path/to/outputDir", myFun);

The result is a graph of computation, and the interesting thing is that
it is reactive. In other words, when an input file changes, the system
is updated incrementally.