New rule: subdirectories have to be created from the direct parent path.
So to create the subdirectory /a/b/ you need to be in /a/.
To create the subdirectory /a/ you need to be in /.
etc ...

This is because lazy directories could create subdirectories.
When this is the case, the presence (or absence) of a subdirectory
ends up depending on the computation of a lazy value.

This is not desireable. Removing a lazy value from the cache should
not lead to any semantic change in the program.

By adding this rule, it is now possible to check if a lazy directory
is in the creation path, and when that is the case, to compute the value
that created the subdirectory.