First implementation of connect/reconnect.
It is now possible to "watch" a directory (does not yet work for
inputs, but soon).

sqlive --connect /path/to/dir --cmd "cmd"

The command cmd will first be called with all the files present
in the directory. It returns a session number on stdout.

After that, every time a change occurs, the command is called again,
this time with only the files that changed.

If the command returns a non-zero status, it becomes disconnect.
You can check the status of everything by calling: sqlive --sessions.

To reconnect a disconnected session: sqlive --reconnect sessionID.

This will first replay everything that was "missed", so all the changes
since the command was disconnected and then reconnect.