First implementation of external pointers.
External pointers are objects that are allocated outside of the
heap. They are not allowed in persistent mode, an error message
is displayed should one attempt to use them in that mode.

Checkout TestExternalPointer for how to use them.
In a nutshell:

SKFS.ExternalPointer(myInteger, myDestructor)

Is an object that will be recognized by the sync function. When
the reference count hits zero, the destructor will be called with
the integer myInteger.