Bug fix on lazy async dirs.
We changed the policy for async directories (after much back and
forth).

We used to recompute the data, check if it had changed or not,
and then dirty the readers of that data (had it changed).

Turns out that this strategy is extremely wasteful when a lot
changes at the same time. You end up spending more time comparing
values than doing actual work.

So we decided to systematically invalidate the transitive closure
of a lazy data that was invalidated. And wait for a call to force
us to recompute that data (shall it be needed).

It turned out to work better in most cases.

Now the problem with that change is that it broke the way async
was working. Async computes a certain amount of pending lazy
chunks defined by the user. It then "dirty" the readers of that
data (because the value has now changed).

However, the async entry that had just been computed was still
considered dirty. Because of that, the update later on removed it
from the up-to-date data.

So we were computing something, only to immediately remove the
result ...

This diff fixes that.