Added a check that the reference count on the context is never greater than 2 when not in concurrent mode.
Currently, the global context is a global accessed through
gContextGlobal. Every time a new context "escapes", the reference
count is incremented by one.

In a normal situation, when no concurrent reader is present,
the ref count should never be higher than 2.

I added a check throwing an exception when that is the case (in
memory mode, when the option --data is absent). Because by
definition, there whould not be any concurrent readers in that
mode.