Fix Float.toString().
The previous implementation was bugged (it used the non-normalized
value cast to int as the integral part, which obviously failed for
cases where the scientific notation would be used).
Moreover, the `printf()` manpage states that a positive exponent
should be prepended with `+` in scientific notation, which was not
implemented.
The current implementation still differs from the result of
`printf("%.17g")` since we store the decimal part in an int32, which
limits the result to 10 significant digits.