Support exceptions in `LazyDir` definition.
Currently, throwing an exception during the definition of a `LazyDir`
value leaves the `LazyDir` in an inconsistent state (the value for
that key is set to `LCycle()` and further requests for that key would
throw a `CyclicExn()`).

This commit catches possible exceptions during definition, cleans up
the state (explicitly setting the value to `LAbsent()` and resetting
the value of `context.reads`) before re-throwing.