Making the build incremental
Right now, we don't use the incremental capabilities of skc at all.
Which is a bit unfortnate.

It's true that the type-checking is fast enough in most use cases
in practice. But re-typing from scratch can take a few seconds,
while we could have those errors back in a few milliseconds.

I really see no reason to wait. Plus, it makes us dogfood things
which I think is nice.

This diff changes the build of build/skdb and build/out32.wasm
to build a "state" file, which is then reused when something
changed.