Got rid of stripdebug
There is no need to compile with the option -static anymore because
we will probably not be shipping a binary anymore.

Back when this was the case, we needed the same binary to work
with any linux x64 platform.

I also removed the stripdebug, which was only useful to make
the binary smaller. Having the debug info at runtime might
actually be useful at some point. It's true that is not super
useful right now with the skip code, because the skip compiler
emits very few debugging info. But we have some C code in there
too, probably worth keeping that info around.