Fixing the build
We merged the privacy layer with the master branch so that
everybody using M1/M2 macbook could build the compiler and start
working.

The problem is that the privacy layer was still in a prototyping
phase, not really tested, code and dependencies not well thought
out. It was still in a phase were I would be playing with the
privacy, and see how it "felt"  without worrying too much about
correctness.

One of the problem is that while I was prototyping, I introduced
a dependency from the SKFS module to SKSQL module, which is really
not desirable long term.

And that breaks the build ...

So I just commented out the offending code, that should only
trigger on a privacy change (and so not really affect anyone for
now).

I suggest that we decide what to do once we know (and test) how
the privacy layer should work.

Feel free to push back if you think that's a bad idea.