Converted run_skdb.js to typescript
We have been hacking on run_skdb.js for long enough. It was useful
to get things going and see how the DB felt. But we need a more
robust working environment.

This diff converts run_skdb.js to typescript, and added a proper
test framework (very rudimentary for now).

This is work in progress. But we got at least enough to convince
ourselves that basic sql queries actually work before pushing
changes.