Fixed jsRoot initialization
We were using "null" to indicate that the root map had not yet
been initialize.

With something like this:
this.roots = null;

And then for initialization, we would do something like this:
if(this.roots === null) {
  this.roots = ...
}

That was working well until I added "strict" mode of typescript.
Typescript complained that "roots" was nullable, so I changed
the code to:
this.roots = new Map();

And of course, forgot to update the initalization code.
Of course non of this would have happen with proper testing, which
I am adding as we speak.