Skargo
Introducing Skargo, the package manager for skip. It is heavily
inspired from Cargo (the Rust package manager), hence the name. It
currently supports build/test with local dependencies.

In order to declare a package, a simple `Skargo.toml` file must be
placed in the project's root. Then, the directory layout is used to
discover the src/tests files (namely, every `.sk` file in `src/` is
assumed to be a source file, and every `.sk` file in `tests/` is
assumed to be part of the unit test sources. The `skargo test`
sub-command builds and executes the test files (using the
`SKTest.main` harness by default).