Decouple the client from server
The client orchestrates the server. This now decouples with a very
basic json-based protocol.

Currently user and password are passed with request to minimise
changes. The db file name is not as this would be fairly coupled. So
for now we assume there is just one db.

This will evolve to have auth info and database associated with the
connection rather than the protocol's messages.