Catch all Exception for wasm mode.
In native mode, each new process of skdb get their own copy of
the globals (because the process is created by the kernel).
So we don't need to clean up anything after ourselves after a
crash.

Things are little bit different in Wasm, given that the program
stays alive forever, if we just crash, the globals will stay
in the state they were in.

Of course we could clean up after ourselves. But that would be
difficult to maintain. A much better approach is to catch all
exceptions (and print an error if need be).