Proper handling of exceptions in Wasm.
We were using a phony exception for errors in the C runtime. As
a result, when one of those exception was raised, the whole
mechanims supposed to catch all exceptions was failing, throwing
another exception. And because of that, we were leaking memory
when too many errors where thrown.

This diff fixes that:
- Created a new type of exceptions called CRuntimeExceptions
- Made sure that we use those everywhere in the C code
- Made sure that we could have 10000 errors in a row and still
  have memory.