Output an in-band checkpoint
This allows a tailer to know what version of the table they've seen up
to.

The checkpoint could occur at any time (might be useful for heartbeats
later on) but should not be output interleaved with a transaction.

It must be in-band, i.e. as part of the tailed output stream, so that
any processor can sequence it reliably.