SKTest
SKTest is a replacement for the current `UnitTest` framework that is
part of the stdlib.
It leverages the `#forEachFunction` macro so that every function
annotated `@test` gets executed by the test harness (`SKTest.main`).

The expectation-style helper functions are quite minimalistic for now,
but it will evolve with our needs.