Refactored all the node logic out
We don't really care about node, we just want to run a couple of
tests with it. So I just factored out all the logic and added a
few dirty hacks to get our tests running.

Since that logic is only seen/used by us that's ok.
If we wanted to add proper support for node, that solution would
not be acceptable.