Don't use a default checkpoint of zero
I thought using 0 instead of None for the checkpoint was more elegant
but it actually has timing issues: it's sensitive to how the data
arrives. You can get spurious zeros being emitted. This shouldn't
cause any actual harm - write csv takes the max - but it's noise that
could be confusing.