Change write-csv to use assignment semantics on row repeat
This writes a repeat of 0 for a deletion. Effectively a tombstone
record. We could optimise and actually remove the row. But we need
tombstones later when we add a clock for the case when a disconnected client
replays inserts that happened before a delete (we want the delete to
win so we need to record a clock against a row with zero repeat).