Save the obstack at the begining of the main.
The obstack was saved to late in Main.sk. As a result, everything
that was called between the beginning of the main function and
the vtry was leaking. That's not a problem in native where the OS
is going to reclaim memory, but it is in wasm, where the main is
a function like any other.