Introduced a type for Names
SQL is case insensitive. Up until now, we were using strings to
represent names and lowercasing when necessary. This was very
error prone and lead to quite a few bugs.

This diff introduces a new type called "Name" where the
lowercasing is performed at the construction of the object.
The Name comparison function is case insensitive (which is what
we want), but error messages and JSON objects still use the
original names (which is also what we want).

The AST now uses that type everywhere.