Re-implement as explicit wipe out and then set
Previously the implementation relied on the source being the same and
the repeat value set so that a single row would be over-written, thus
achieving assignment semantics.

This causes the problem of concurrent inserts clobbering each other
with a classic data race. Now we restore the unique sources to avoid
that and explicitly wipe out any previous rows then insert the new
ones.