First implementation of incremental sharding
When joining two tables, what can happen is that one side has many
more elements than the other.
When that happens, changing the side that already has a lot of
elements is costly, because we have to unnecessarily reconstruct
all the elements involved.

The solution was to introduce sharding: the side with fewer
elements emits more than one key and the side with less elements
just choses one of the keys at random. This reduces the number of
elements that have to be reconstructed on a change, and therefore
the overall cost.

Concretely:

Row1: 1, 'a'  ==> (1, 'a'), (1, 'hello') <==|
Row2: 1, 'b'  ==> (1, 'b'), (1, 'hello') <==| Row1: 1, 'hello'
Row3: 1, 'c'  ==> (1, 'c'), (1, 'hello') <==|

In this example the left side has more elements projecting to 1
that the right side. So instead of projecting to 1, we project
to 1-x where x is a sharded number (let's say 4).

Row1: 1-3, 'a'  ==> (1, 'a'), (1, 'hello') <== Row1-0: 1, 'hello'
Row2: 1-1, 'b'  ==> (1, 'b'), (1, 'hello') <== Row1-1: 1, 'hello'
Row3: 1-2, 'c'  ==> (1, 'c'), (1, 'hello') <== Row1-2: 1, 'hello'
                                           <== Row1-3: 1, 'hello'

Notice that the shard on the left side is chosen at random.
With this technique, changing (or adding) an element on the left
side is now a lot cheaper because there are way less elements
involved.

But one last problem remains, what shard should we chose?
Up until now, we were using a scheme based on the sizes of the
tables at the time the join was created. Although that often works
well, it breaks appart whenever the tables and views are first
created and the elements later inserted. This seemed like a
contrived example, but it's actually useful client-side, when one
is trying to load data one by one, to avoid blocking the database.

So we now changed that scheme to compute stats on the elements
themselves on the fly. The implementation is not very clean with
respect to SKFS because we are modifying a directory while we
are computing an update (which is normally considered
bad-practice). But hey! It works! ;)