Fixed setLocalContext for jsRoots
JSroots to be computed need access to the context. The problem is
that exposing the SKFS API to the JS world seemed a little bit
too complicated. So instead, what we do is that we put the context
in a global automatically when needed.
The problem is that the logic adding/removing the so called local
context (maintaining the global) was wrong. This diff fixes that.