Fix js reload
There are currently two states for skdb on JS. The state that
lives in the wasm, and the state that lives in the SKDB JS object.

When the database is reloaded, the wasm state is reinstored, by
reading from indexedDB. But not the JS state, that is rebuilt
from scratch.

The problem is that the registerFun mechanism was assuming that
the JS state was the same as at creation time. And this was
leading to duplicate callback triggers.

This diff cleans that up, the JS hooks are reinitialized on the
wasm end, to reflect the fact that the JS state was restarted.