Changed exceptions default message.
The default message for exceptions was not very informative. It
was only giving the class name of the exception. Using the
inspect method, we should be able to see a lot more information.

This is only for debugging purposes since no exception should be
thrown as is to the user anyway.