Commit the write-csv txn on receiving the checkpoint
The checkpoint acts as a nice end of txn marker. Use it to signal
commit. Previously we waited until stdin had nothing left, but this is
pretty flaky.

- network delay could cause inappropriate commit

- a stream that is cut off partially could be committed

- the test was based on using select to see if the fd was readable.
  buffering can affect this and show it as not having data when in
  fact there is. this problem was actually happening with the server
  pipes.

Aside from there being theoretical issues with this, we were actually
preventing duplicates from being written. What happens is the
duplicate is sent up as multiple rows, which are applied as separate
txns (due to buffering) and these clobber each other rather than build
up.

As a nice side benefit to this change we reduce syscalls in a
potentially hot loop, which might make this marginally faster. ;)