Added the npm target to the Makefile
This target is the one we use before exporting the npm package.
First we copy over the file to the skdb repo (using exportNPM.sh),
then we publish the package.

But before we do that we need to make sure that all the wasm files
and ts files have been built properly, which is what this target
is for.