Implementation of tomb collector
We use to just accumulate tombs. This is wrong for two reasons: one, we are leaking memory, two we are getting slower and slower on data that changes a lot.
In this diff, we implement a "tomb collector", that is, an algorithm that collects tombs that are sufficiently old.

To make this works, we need to define a point in time (in the past) where diffing is forbidden.
That point is called the tombLimit and it's reflected in the API of getChangesAfter.
If getChangesAfter returns true, that means that we are dealing with a reset, and that all
the elements that are not present in the return set should be considered dead.