INSERT OR REPLACE speedup (and fix)
We use to only remove elements involved in an insert or replace
when these elements where creating conflicts with "primary key"
or "unique" annotations.

However, in the presence of a unique index, the annotation was
ignored (the feature was not supported).

This diff fixes that. Unique indexes are no longer ignored, and
we make sure that insert of replace removes elements involved in
any conflict.

This is a necessary step to support primary key based on multiple
columns.