Don't special case 0 for resets.
We use to special case diffs from 0 to never trigger a reset.
However, this could lead to problems with the server sync.

Copy pasting the comment made in PR #165 for context:

"The server streams down all the data and finally it's tick value,
which is used for re-establishing the tail. I think today we won't
write anything until we see that value. But if we changed it so that
we wrote optimistically - maybe to avoid holding the whole state in
memory? - then we may be forced to re-establish from zero if we lost
the connection without receiving the tick. We'd need the js to
understand what had happened and wipe out the data. This is maybe a
bit hypothetical, but it's what I have in mind when I think we
shouldn't special case zero."