Relaxing JS sync
Synchronization to disk is slowing down the main thread when a lot of
pages need to be synced. This is unacceptable, the UI always need to
have the main thread avaible.

The way it used to work:
- An SQL transaction just finished
- At the end of the transaction the pages were copied (synchronously)
- The IndexedDB transaction was started (asynchrounously)
- We were careful not to overload indexededDB, meaning, not to start a
  new transaction when an old one did not finish.

The way it works today:
- An SQL transaction just started, we bump up the transaction number,
  this will interrupt any existing sync that was happening
- The pages are copied (asynchronously), and we check that no new
  transaction has started between each page copy.
-- If the transaction succeeded, nothing to do
-- If the transaction was interruped, we restart the process from
   scratch.

What's nice is that in this new version, there is zero work done on
the main thread. The only downside is that if we keep on interruping
transactions we could be losing data.  But that's an ok compromise.