Fix of the dirtyPages clearing algorithm
The logic clearing the dirtyPages was wrong. It was clearing the
dirtyPages when the indexedDB transaction was successful but
without checking that another transaction actually started.

So we could end up in a scenario where the finishing transaction
was wiping out the pages of a queued up transaction.

To fix this, we only wipe out the dirtyPages when there are no
transaction waiting.