First implementation of alternative SQL engine
The problem: our current SQL engine is based of SKFS, which is great
for reactive queries. The problem is that we also use the same engine
for non-reactive queries, which can result into a significant slow
down. Imagine the scenario where you are running a single select, why
would it need to be reactive? The context is immutable, we could run
that on the fly without any problem.

In this diff, we introduce the first version of SQL engine that walks
the data in SKFS, but without trying to be reactive.