[runtime] Return subprocess termination status on non-exit termination
If the subprocess spawned by System.subprocess terminates for a reason
other than exiting, such as via a signal or trapping, then the
returned exitcode is 0. This conflates success and some forms of
failure but not others.

One option would be to enrich System.CompletedProcess with a
termination status as well as an exit code. This might be a good idea,
but the goal of System might not be to mimic POSIX.

A workaround is to return the termination status unchanged unless the
subprocess exited. This will produce a non-zero exitcode value when
the subprocess terminates for a reason other that exiting
successfully.

With this change, skargo no longer successfully terminates silently
when the compiler crashes.