[skargo] Interpret paths in deps relative to their base
Currently when the paths of dependencies are collected, they are kept
exactly as they appear in the Skargo.toml file. This has the effect of
interpreting paths relative with respect to the base directory of the
initial Skargo.toml file, which will not necessarily be correct for
the dependencies.

For example:
>> pwd
/skfs/josh/toy
>> grep path Skargo.toml
std = { "path" = "../../prelude" }
>> grep path ../../prelude/Skargo.toml
skfs = { "path" = "../skfs" }  # Temporary
>> grep path ../skfs/Skargo.toml
grep: ../skfs/Skargo.toml: No such file or directory

Here, instead we want:

>> grep path ../../prelude/../skfs/Skargo.toml
sktest = { "path" = "../sktest" }  # TODO: Move this to dev-dependencies.

This diff implements this by passing the base directory containing the
manifest file when reading the dependencies, prepending it onto the
path found in the dependency's manifest file.