Saving a map when an index preserves the directory order.
Every directory has an order, the same one as the one defined at
creation time in its schema. What can happen very often is that the
index we are trying to create is already there, because that's the
order the SKFS directory was already using. In this diff we optimize
that away, by not introducing an extra directory.