Implementation of --format=js option
When SKDB is used in JS, we spend quite a bit of time pretty-printing
and then parsing JSON objects. This is a waste of time, we could just
build the objects directly from within the DB. That's what the new
option --format=js does. Instead of "writing" to stdout (which in JS
consists in appending strings to the field called 'stdout' of the
runtime object), we append objects to a new field called
stdout_objects. At the end of the query, we check if there was no
error messages, and if that's the case, we return the JS objects
directly.

This is a 2X gain on a typical JS query.