Reformulate FixedDir.getIter etc using an Iterator of found elements
Currently the first element that compares equal to the key is found,
followed by a linear scan forward until a non-equal element is
encountered. This commit adds a function to iterate all the elements
that compare equal to the key, which simplifies getIter, getArray, and
getIterSourceKey, making it clearer what they are actually doing.