Changed the persistence model for JS
SKDB use to be persistent by default, this is no longer the case. One
now needs to call SKDB with a DB name at creation time. If no name is
given, it is assumed that no persitence is needed. The other thing
that changes is that the page saving is no longer called
automatically. It needs to be explicitely called (using the method
"save").