Inlining wasm
The current situation to load wasm files included in a package is
... well ... less than ideal. Packagers disagree with each-other,
nothing works, is a pretty good summary. So it's best if we just
inline the wasm (using a base64 encoded string) ourselves. It's not
great to have it in the type-script file (would have been better to
directly have it in the JS), but that would require revisiting all the
code that produces an npm package, and I would like to take things in
steps.