Create a WASM loading strategy
We try to be smart about the context. We use inlined b64 bytes if
they're there. Otherwise we load using fetch. This defaults to a
mechanism that should work with a packager like Vite. When we build
the node js file (in the Makefile) we replace the fetch line to use
the fs module.