Erroring a socket now just sends the goaway
Previously we would send reset on each active stream and then finally
the goaway. Now we trigger callbacks as if the other side had errored
the stream. This is more efficient on the network and makes the
abstraction easier to use: you can now trigger all stream callbacks by
erroring the socket, which means you don't have to track active streams
yourself.

We do not do this for closing as it is important to be able to close
the socket (which half closes it) and then get the onClose callback
when the other side gracefully closes.