Allocate more space to ensure we can encode to utf8
Although utf8 (previously) defined up to 6 bytes per code point, I was
multiplying by 3. This was based on an old memory, that unicode only
needed 3 bytes to encode it. Modern unicode actually needs 4, and it
looks like utf8 now actually only specs up to 4 bytes in accordance.

And I just pre-allocate as everywhere we do this the strings are
expected to be pretty small. KISS.

I have it in the backlog to add request size limits so we have a hard
bound on what will be allocated.

If allocation is ever expensive enough to show up (these allocs aren't
on the common paths) then we can always do a pass to figure out exactly how
much to alloc. Or do some sort of streaming encoding in to a linked
list of arrays.