Encode device uuid in to auth msg
We lost this when we moved from json to the binary muxed socket.

It wasn't implemented because it's not actually used any more. But I
figure this is really useful information that we should keep and have
in the protocol from the beginning. We can always zero it out and not
send in the future.

Potential uses are per-device analytics, understanding reconnects and
making them more efficient, etc.