Fix: only subing roots on reboot means we do not learn about past roots
The subscribe picks up where it left off. We need a new subscription
that gives us the past data in order to rebuild state.

This quick fix does however leak subs, which is probably why the logic
was there.

I removed RebootStatus class as it's now unused but we may want to
re-introduce with a better fix.