Change delta output to include optional destination time
This represents the time at destination - where this diff will be
applied - that we are in sync up to. In other words: this defines the
snapshot of the destination data that this change should be applied
to. It should NOT be applied to data newer than this, as the producer
of the diff has not seen that data and hasn't expressed a change to
it.