Support for filter while tailing
You can now write: skdb tail tableName 'a > 10'

This is a necessary first step to support mirrorTable with a
filter. Originally, I considered adding arbitrary support for SQL, but
realized that 99% of the time, people will just want to filter on the
table being mirrored directly and we could support that for much
cheaper. Also, the user can create a virtual view and get whatever
data they want.