Add a check to see if a table's schema can be mirrored
We don't currently support unique constraints and likely in the future
we'll want to limit other things.

Note on implementation: this was done entierely within skdb rather
than the server propagating this information. It has the downside that
an unsupported table gets created and isn't cleaned up (once we have
drop we can do this).

But the main reason is so that we don't need to understand in the
server what version the client is running and track capabilities. i.e.
this prevents old clients from mirroring things they cannot support.

It also is a lot less work than adding the query to the replication
protocol - but that's just a bonus. ;)