Backoff before first retry
I initially didn't add this as I figured an immediate retry and then
backoff (in the loop) would work well. But there's a gap.

If the server accepts the connection, but then fails and errors the
socket, this spins as fast as it can in a reconnect loop.

This behaviour was easily reproduced if you connected to the server to
access a database that doesn't exist. The websocket handler accepts
the connection, but then throws before establishing the MuxedSocket.
This causes the JS to infinitely bounce between replaceFailedSocket
and attachSocket.