Direct mode speedup
There is a mode were we know that the query does not need to be
reactive. It's when a user is running a readonly query, outside of a
transaction. In that mode, we run the so-called "direct mode". It's a
mode where the evaluator prints results directly without ever building
a directory in SKFS. This diff speeds things up by taking control over
the loop used to iterate and therefore saving a lot of overhead. The
idea is to be faster in the common case.