First implementation of new privacy framework
We use to have a more complex privacy framework, where sql queries
could be used to define groups. However, it was a bit complicated. The
new scheme is much simpler. There is a table called
skdb_group_permissions, with the following schema:

groupID INTEGER, userID INTEGER, permissions INTEGER

Where permission is a bit mask. The first bit (& 1) is set to 1 if the
user can delete. The second bit (& 2) is set to 1 if the user can
insert, the third (& 4) if the user can read.

The user NULL is special, it is the default for all users. If NULL is
not specificied, then all users absent from the group will have no
access (the default value is 0).

To use a group, you just need to place its ID (an integer) in a column
named skdb_groups. That column can be added to any table. When that's
the case, the DB will enforce the policy.

There are two other tables, skdb_table_permissions and
skdb_user_permissions that lets one control permissions at the table
level and at the user level.

Privacy policies are only enforced at the level of tailers and
writers. That's because supporting it in the SQL layer would be too
expensive. All SQL queries are therefore interpreted as root.