Removed INSERT OR REPLACE optimization
There was an optimization in place that consisted in preemptively
removing rows from a table on a REPLACE OR INSERT. This optimization
was a win for small table but turned out to be a huge regression for
larger ones. The reason: indexes do not work on tables passed the
first write statement ... And that needs to be fixed regardeless
obviously.

Turning this off for now, we can revisit later.