Fix broken wasm loading using node from npm
This makes loading from wasm using node much cleaner. We have a single
mechanism that should work everywhere now. It also fixes the cli not
working from the npm package. After this lands and we ship an npm pkg,
you should just be able to do `npx skdb` to get a repl.

We didn't do this previously because the tests were constructed by
cating files together and piping them in node directly to run. This
means the source wasn't backed by a file and so import.meta.url was
undefined. Now that we use files everywhere, this should work.