Refactored index table to be based on table names
Indexes were found using SKFS directory names, and the problem with
that scheme was that we could have multiple versions of the same index
(cf SqlTable.sk). When that happened, we had to do a mangling where we
would have to pretend that the table was using the original directory
name. This diff fixes that. The indexes are now found using the table
name (whatever the version).