Optimized noop select (useful for joins)
Whenever we write a typical join:
select * from t1, t2 where t1.x = t2.x

Very often, the result of the merge is exactly the result of the
select. In this optimization, we rename the merge directory name to
match the one of the select and turn the select into a noop.