Check the results of native-test
Currently `make native-test` runs the native tests, but does not check
the results. Since this is what backs the CI job, the CI jobs will be
green whether or not the tests pass.

This PR adds some result checking to the `make native-test`, so that
the exit code will be non-zero if the output includes non-passing
results. This is a fragile grepping of stdout implementation, but only
requires a small change.

In order to make the tests pass:

  - sql/test/memory/run.sh is changed to make grepping for pass/fail
    easier