Fix gradle hang issues by building java with native arch
Gradle had some issues forking processes with the emulation (rosetta
or qemu) so we do the java build natively. This makes things faster
too.

I also took the opportunity to use an openjdk image from eclipse
rather than messing around installing sdkman which is not appropriate
for a prod build.