Disable OuterIst locals optimization pass
A long time ago, skip had multiple backends. One back-end in
particular was the javascript backend. This backend was very useful
before we got the native backend, since we were bootstrapping, using
the interpreter was egregiously slow. To make sure this backend was
reasonbly fast, we implemented a bunch of optimizations back in the
days. However, those optimizations are not very useful anymore because
the backend does all of that (and much more!). So I just disabled the
optimization pass that was done at the OuterIst level.

This is a fix for #234