Do not actually remove directories
If a purge occurs during a transaction, the resolveContext
implementation detects it, issues an error message, and returns the
original context unchanged, dropping the changes. This can happen in
particular if a client is trying to be active but is starved due to
scheduling.

The code that this check avoids does detect such resets and handles
the changes within a directory correctly.

If a directory is removed and a reset occurs, however, the removed
directory is not detected as having been changed, and in particular it
will not be deleted.

This issue with directories is not specific to context resolution,
however. It appears that tailers and everything else is just as
susceptible to this.

This PR proposes to work around / delay fixing this by only recording
that directories have been deleted, but not actually deleting them,
and also letting context resolution proceed when resets happen. This
potentially leaks memory since the directory maps could grow over time
if many directories are removed.

This is enough to get the current tests to pass reliably.

It seems that a proper solution to this would require a mechanism
wherever directory changes are requested, to enumerate the full set of
current directories when a reset happens, so that clients can remove
deleted directories they still have.