Control the timing of tail by running on demand
By tracking the output and tailing incrementally, it simplifies things
a lot. We can now control exactly when output is generated and there's
no chance of race conditions. The cost is encoding an understanding of
the serialisation, but who cares? This is supposed to model the system
and understand how it works.