Refactor and optimize conversion of Skip to C strings
Introduce a `sk2c_string` function that the rest of the runtime uses
to convert Skip strings to C before passing them to C libraries or
system calls.

This function tests is the skip string is already nul-terminated, and
if not copies it adding a nul terminator. The string is checked for
embedded nul characters, and exits if found.

While doing this, I noticed that `SKIP_getenv` and `SKIP_unix_open`
were leaking the string copies.