Store compiled instead of raw select in SelectFile
Currently SelectFile carries a non-compiled Select, which is compiled
on both the write and read side of virtual views. This PR changes this
to store the compiled select in the viewsDir, that is computed on the
write side already, and avoids recompiling it when it is read out of
viewsDir. This also has the result of removing many uses of the
Options which were only passed around to recompile the select.