Add test for parameterized statements
Checks a range of value data types that can be passed as parameters.

Checks parameters in `insert values` position, parsed as "values", and
in `select where` position, parsed as "expressions". Also minimal
testing of parameters in `where`-clauses in `delete` and `update`
statements, and in `set` clauses of `update` statements.