Improve expectation definition
Check that the result set has converged. This was an unintended gap
before. The checks were intended to ensure that all results were the
same but perhaps one of a set. But it was too loose and just checked
that each was from the set.

We also capture the peer that is out of line and output it for easier
debugging.