Fix reset test to correctly fail
The new privacy framework passes a function that always returns true
to reset when testing if the row is visible to the user. This should
be a bug but all tests were passing. This commit fixes this. The test
had bitrotted as the reset logic had evolved. The 'wasSeen' test was
accidentally causing the test to pass.