Add support for query parameters to JS api
This PR adds an argument to the `runLocal` function that wraps the
call to `skip_main`. This argument is a map from parameter key strings
to value strings, which defaults to empty. In case the params map is
non-empty, the call to `skip_main` is adjusted by adding the
`--expect-query-params` flag to the `skdb` command line, and the
params map is encoded into a json map and prefixed to stdin, as
expected by skdb.

This PR modifies the callers of `runLocal`: `cmd`, `sqlRaw`, `sql`,
and `insert`. For the first 3, a new argument for the params is added
and just passed along to `runLocal`. For `insert`, the provided values
are used to build a params map using generated key names, so that the
user-provided values do not appear in an SQL string.